requirements.txt
streamlit
pillow
google-generativeai
python-dotenv
webcolors
PIL 
io
base64
colorsys
random  